+++ resources/js/pages/product/index.js
@@ -1,6 +1,7 @@
 import { __ } from '@wordpress/i18n'
 import { Link } from 'react-router-dom'
 import { Container, Table, Button, Alert } from 'react-bootstrap'
+import Swal from 'sweetalert2/dist/sweetalert2.js'
 import { DeleteProduct } from '../../http/product'
 import { useEffect, useState } from '@wordpress/element'
 const { useSelect, dispatch } = wp.data
@@ -18,12 +19,22 @@ export const ProductList = () => {
     }, [productList])
 
     const deleteProduct = (productId) => {
-        DeleteProduct(productId).then((response) => {
-            dispatch('smartpay/products').deleteProduct(productId)
-            setResponse({
-                type: 'success',
-                message: __(response.message, 'smartpay'),
-            })
+        Swal.fire({
+            title: __('Are you sure?', 'smartpay'),
+            text: __("You won't be able to revert this!", 'smartpay'),
+            icon: 'warning',
+            confirmButtonText: __('Yes', 'smartpay'),
+            showCancelButton: true,
+        }).then((result) => {
+            if (result.isConfirmed) {
+                DeleteProduct(productId).then((response) => {
+                    dispatch('smartpay/products').deleteProduct(productId)
+                    setResponse({
+                        type: 'success',
+                        message: __(response.message, 'smartpay'),
+                    })
+                })
+            }
         })
     }
